+++ src/apps/gofumpt.rs
@@ -1,5 +1,5 @@
 use crate::apps::{GetRTACmdArgs, get_rta_command};
-use crate::config::{Application, Applications, Config};
+use crate::config::{Application, Applications, Config, Operation};
 use crate::domain::{DetectedStack, EnabledWhen, Fix, Result, Tool};
 use big_s::S;
 use std::fmt::Display;
@@ -28,7 +28,7 @@ impl Fix for Gofumpt {
         stack: &DetectedStack,
         config: &Config,
     ) -> Result<Vec<conc::Executable>> {
-        let ignores = config.ignores_for_app(self)?;
+        let ignores = config.ignores_for(self, Operation::Fix)?;
         let files = stack.files.remove(&ignores);
         if files.is_empty() {
             return Ok(vec![]);